Allow i64 type to be used in JSON in more places, mainly for time values and file sizes.
FossilOrigin-Name: f89a32d7820a13b5b82245d583b29e028b1efcf2

+++ src/json.c
@@ -269,7 +269,7 @@ cson_value * json_new_string_f( char const * fmt, ... ){
   return v;
 }
 
-cson_value * json_new_int( int v ){
+cson_value * json_new_int( i64 v ){
   return cson_value_new_integer((cson_int_t)v);
 }
